Refactor copybook definition module
